+++ src/main/java/edu/illinois/library/cantaloupe/resource/RequestContextMap.java
@@ -29,6 +29,7 @@ final class RequestContextMap<K, V> implements Map<K, V> {
     static final String FULL_SIZE_KEY        = "full_size";
     static final String IDENTIFIER_KEY       = "identifier";
     static final String LOCAL_URI_KEY        = "local_uri";
+    static final String PAGE_COUNT_KEY       = "page_count";
     static final String PAGE_NUMBER_KEY      = "page_number";
     static final String METADATA_KEY         = "metadata";
     static final String OPERATIONS_KEY       = "operations";
@@ -95,6 +96,8 @@ final class RequestContextMap<K, V> implements Map<K, V> {
                 Format format = backingContext.getOutputFormat();
                 return (format != null) ?
                         (V) format.getPreferredMediaType().toString() : null;
+            case PAGE_COUNT_KEY:
+                return (V) backingContext.getPageCount();
             case PAGE_NUMBER_KEY:
                 return (V) backingContext.getPageNumber();
             case REQUEST_HEADERS_KEY:
@@ -158,6 +161,7 @@ final class RequestContextMap<K, V> implements Map<K, V> {
         size += (backingContext.getMetadata() != null)        ? 1 : 0;
         size += (backingContext.getOperationList() != null)   ? 1 : 0;
         size += (backingContext.getOutputFormat() != null)    ? 1 : 0;
+        size += (backingContext.getPageCount() != null)       ? 1 : 0;
         size += (backingContext.getPageNumber() != null)      ? 1 : 0;
         size += (backingContext.getRequestHeaders() != null)  ? 1 : 0;
         size += (backingContext.getRequestURI() != null)      ? 1 : 0;